Initial refactoring of Login component
