+++ src/components/Login/Login.js
@@ -5,10 +5,10 @@ import agent from '../../agent';
 import {connect} from 'react-redux';
 import {
   UPDATE_FIELD_AUTH,
-  REGISTER,
-  REGISTER_PAGE_UNLOADED
+  LOGIN,
+  LOGIN_PAGE_UNLOADED
 } from '../../constants/actionTypes';
-import registerStyles from './Register.module.css';
+import loginStyles from './Login.module.css';
 import DialogPage from '../common/DialogPage/DialogPage';
 import Form from '../common/Form/Form';
 import FormButtons from '../common/FormButtons/FormButtons';
@@ -22,26 +22,21 @@ const mapDispatchToProps = dispatch => ({
     dispatch({type: UPDATE_FIELD_AUTH, key: 'email', value}),
   onChangePassword: value =>
     dispatch({type: UPDATE_FIELD_AUTH, key: 'password', value}),
-  onChangeUsername: value =>
-    dispatch({type: UPDATE_FIELD_AUTH, key: 'username', value}),
-  onSubmit: (username, email, password) => {
-    const payload = agent.Auth.register(username, email, password);
-    dispatch({type: REGISTER, payload})
-  },
+  onSubmit: (email, password) =>
+    dispatch({type: LOGIN, payload: agent.Auth.login(email, password)}),
   onUnload: () =>
-    dispatch({type: REGISTER_PAGE_UNLOADED})
+    dispatch({type: LOGIN_PAGE_UNLOADED})
 });
 
-class Register extends React.Component {
+class Login extends React.Component {
   constructor(props) {
     super(props);
     this.changeEmail = ev => this.props.onChangeEmail(ev.target.value);
     this.changePassword = ev => this.props.onChangePassword(ev.target.value);
-    this.changeUsername = ev => this.props.onChangeUsername(ev.target.value);
-    this.submitForm = (username, email, password) => ev => {
+    this.submitForm = (email, password) => ev => {
       ev.preventDefault();
-      this.props.onSubmit(username, email, password);
-    }
+      this.props.onSubmit(email, password);
+    };
   }
 
   componentWillUnmount() {
@@ -51,36 +46,28 @@ class Register extends React.Component {
   render() {
     const email = this.props.email;
     const password = this.props.password;
-    const username = this.props.username;
-
     return (
-      <DialogPage title={'Зарегистрироваться'}>
+      <DialogPage title={'Войти'}>
 
-        <div className={registerStyles.linkRow}>
-          <Link to="/login">
-            <p className={registerStyles.link}>Уже есть аккаунт?</p>
+        <div className={loginStyles.linkRow}>
+          <Link to="/register">
+            <p className={loginStyles.link}>Хотите создать аккаунт?</p>
           </Link>
         </div>
 
         <ListErrors errors={this.props.errors}/>
 
-        <Form onSubmit={this.submitForm(username, email, password)}>
-
-          <InputText
-            label="Имя пользователя"
-            placeholder="Username"
-            value={username}
-            onChange={this.changeUsername}/>
+        <Form onSubmit={this.submitForm(email, password)}>
 
           <InputText
             label="E-mail"
-            placeholder="Username@nomoreparties.space"
+            placeholder="E-mail"
             value={email}
             onChange={this.changeEmail}/>
 
           <InputPassword
             label="Пароль"
-            placeholder="********"
+            placeholder="Пароль"
             value={password}
             onChange={this.changePassword}/>
 
@@ -89,14 +76,15 @@ class Register extends React.Component {
               className="btn btn-lg btn-primary pull-xs-right"
               type="submit"
               disabled={this.props.inProgress}>
-              Зарегистрироваться
+              Войти
             </button>
           </FormButtons>
 
         </Form>
+
       </DialogPage>
     );
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Register);
+export default connect(mapStateToProps, mapDispatchToProps)(Login);
