+++ src/components/common/DialogPage/DialogPage.module.css
@@ -1,0 +1,18 @@
+.page {
+  display: flex;
+  justify-content: center;
+  padding-top: 32px;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  flex: 1;
+  max-width: 540px;
+}
+
+.title {
+  composes: headerTextMedium from "../common.module.css";
+  margin: 16px 0;
+}
