+++ src/components/common/Form/Form.module.css
@@ -1,6 +1,6 @@
 .form {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  width: 100%;
 }
