+++ src/components/common/FormButtons/FormButtons.js
@@ -3,18 +3,18 @@ import React from "react";
 import formButtonsStyles from './FormButtons.module.css';
 
 function FormButtons({children}) {
-    return (
-        <div className={formButtonsStyles.formButtons}>
-            {children}
-        </div>
-    );
+  return (
+    <div className={formButtonsStyles.formButtons}>
+      {children}
+    </div>
+  );
 }
 
 FormButtons.propTypes = {
-    children: PropTypes.oneOfType([
-        PropTypes.element,
-        PropTypes.arrayOf(PropTypes.element),
-    ]).isRequired,
+  children: PropTypes.oneOfType([
+    PropTypes.element,
+    PropTypes.arrayOf(PropTypes.element),
+  ]).isRequired,
 };
 
 export default FormButtons;
