+++ src/components/common/common.module.css
@@ -1,54 +1,54 @@
 .headerTextLarge {
-    font-family: Neue Machina, Arial, sans-serif;
-    font-style: normal;
-    font-weight: 800;
-    font-size: 64px;
-    line-height: 64px;
-    color: var(--text-color);
+  font-family: Neue Machina, Arial, sans-serif;
+  font-style: normal;
+  font-weight: 800;
+  font-size: 64px;
+  line-height: 64px;
+  color: var(--text-color);
 }
 
 .headerTextMedium {
-    font-family: Neue Machina, Arial, sans-serif;
-    font-style: normal;
-    font-weight: 800;
-    font-size: 36px;
-    line-height: 40px;
-    color: var(--text-color);
+  font-family: Neue Machina, Arial, sans-serif;
+  font-style: normal;
+  font-weight: 800;
+  font-size: 36px;
+  line-height: 40px;
+  color: var(--text-color);
 }
 
 .headerTextDefault {
-    font-family: Neue Machina, Arial, sans-serif;
-    font-style: normal;
-    font-weight: 800;
-    font-size: 24px;
-    line-height: 28px;
-    color: var(--text-color);
+  font-family: Neue Machina, Arial, sans-serif;
+  font-style: normal;
+  font-weight: 800;
+  font-size: 24px;
+  line-height: 28px;
+  color: var(--text-color);
 }
 
 .text {
-    font-family: Golos, Arial, sans-serif;
-    font-style: normal;
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 24px;
-    color: var(--text-color);
+  font-family: Golos, Arial, sans-serif;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 24px;
+  color: var(--text-color);
 }
 
 .linkText {
-    font-family: Golos, Arial, sans-serif;
-    font-style: normal;
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 24px;
-    text-decoration-line: underline;
-    color: var(--link-text-color);
+  font-family: Golos, Arial, sans-serif;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 24px;
+  text-decoration-line: underline;
+  color: var(--link-text-color);
 }
 
 .caption {
-    font-family: Golos, Arial, sans-serif;
-    font-style: normal;
-    font-weight: 500;
-    font-size: 12px;
-    line-height: 16px;
-    color: var(--text-color);
+  font-family: Golos, Arial, sans-serif;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 16px;
+  color: var(--text-color);
 }
